40 - improve the directory use
