--- src/app/components/RandomRecipe/SingleRecipe.css
@@ -1,107 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Noto+Sans:wght@400;500&display=swap');
-
-h2 {
-  font-family: 'Ubuntu', sans-serif;
-  font-size: 28px;
-  font-weight: lighter;
-  padding: 50px 118px;
-}
-
-h2 a {
-  font-weight: bold;
-}
-
-.singleRecipe {
-  background-color: #ffffff;
-  border: none;
-}
-
-.fontStyle {
-  font-family: 'Ubuntu';
-  font-style: normal;
-  font-weight: 500;
-  font-size: 30px;
-  line-height: 46px;
-  text-align: center;
-}
-
-.dishName {
-  margin-top: 100px;
-}
-
-.ingredientName {
-  margin-bottom: 30px;
-}
-
-.dishInfo {
-  display: flex;
-  margin-top: 50px;
-  margin-bottom: 50px;
-  justify-content: center;
-  gap: 100px;
-}
-
-.dishImage {
-  text-align: center;
-  margin: 0 30%;
-}
-
-.ingStyle {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  margin: 30px 30%;
-  column-gap: 2rem;
-}
-
-.insStyle {
-  margin: 30px 30%;
-}
-
-li {
-  margin-bottom: 10px;
-  font-family: 'Ubuntu';
-  font-style: normal;
-  font-weight: 300;
-  font-size: 15px;
-}
-
-.rateSection {
-  display: flex;
-  margin: 30px 30%;
-  gap: 40%;
-}
-.rating {
-  border: 1px solid;
-  padding: 2%;
-  background-color: #f0eccf;
-  display: flex;
-  gap: 10px;
-  text-align: center;
-}
-
-@media only screen and (max-width: 600px) {
-  .ingStyle {
-    display: block;
-  }
-  .dishImage {
-    height: 300px;
-    width: 300px;
-    margin: 0 10%;
-  }
-  .dishInfo {
-    display: block;
-    margin: 50px 30%;
-  }
-  .ingStyle {
-    margin: 30px 10%;
-  }
-  .insStyle {
-    margin: 30px 10%;
-  }
-  .rateSection {
-    display: block;
-  }
-  .rating {
-    margin-bottom: 20px;
-  }
-}
