--- src/app/components/RandomRecipe/page.css
@@ -1,15 +1,0 @@
-* {
-  box-sizing: border-box;
-  padding: 0px;
-  margin: 0px;
-  font-family: 'Ubuntu';
-}
-
-.main {
-  background-color: white;
-  color: black;
-}
-
-.filter-showcase {
-  display: flex;
-}
