+++ src/app/components/RandomRecipe/page.js
@@ -1,10 +1,10 @@
 'use client';
 
-import './page.css';
+import '../../page.css';
 
 // import { useSearchParams } from "next/navigation";
 
-import SingleRecipe from './SingleRecipe';
+import {SingleRecipe} from './SingleRecipe';
 
 export default function Home() {
   return (
