--- src/app/components/terms/terms.js
@@ -1,165 +1,0 @@
-import React from "react";
-import Image from 'next/image';
-import './terms.css';
-import bannerTerms from './bannerTerms.jpeg'
-import Link from "next/link";
-
-export const TermsOfUse = () => {
-    return (
-       <React.Fragment>
-        <div>
-        <h2><Link href="#">HOME</Link> {">"} Terms of use</h2>
-    </div>
-    <div className="banner">
-        <Image src={bannerTerms} alt="banner" />
-    </div>
-    <div><h1>Terms of use</h1></div>
-    <div>
-        <p>Create March 27, 2023
-            <br></br>
-            <br></br>
-            <br></br>
-
-1. Overview
-<br></br>
-            <br></br>
-Access to and use of the “What can I cook” site is subject to these terms and conditions of use (“Terms of Use”).
-“What can I cook” site shall include any information or services made available, regardless of the medium. We reserve the right at any time, and from time to time, to modify, suspend or discontinue (temporarily or permanently) the site, or any part of the site without notice.
-We make no claims that the “What can I cook” site or any site’s content is accessible or appropriate in your country.  If you access the site, you do so on your own initiative and are responsible for compliance with local laws and regulations.
-<br></br>
-            <br></br>
-            <br></br>
-
-2. Disclaimer
-<br></br>
-            <br></br>
-The information available through the “What can I cook” site is provided solely for informational purposes on an “as is” basis at user’s sole risk. We makes no guarantees as to the accurateness, quality, or completeness of the information and shall not be responsible or liable for any errors, omissions, or inaccuracies in the information or for any user’s reliance on the information. Users are solely responsible for verifying the information as being appropriate for user’s personal use.
-<br></br>
-            <br></br>
-            <br></br>
-
-3. Our Right to Modify These Terms of Use
-<br></br>
-            <br></br>
-We reserve the right to change these Terms of Use at any time. You should check this page regularly. The changes will appear on the Sites and will be effective when we post the changes. Your continued use of the Sites means you agree to and accept the changes.
-<br></br>
-            <br></br>
-            <br></br>
-
-4. Our Privacy Policy
-<br></br>
-            <br></br>
-This Privacy Policy only applies to properties we own and operate and does not cover any information collected via any other website, application, or service. We are not responsible for the privacy policies or privacy practices of third parties. By using the Sites, you explicitly agree to the terms of this Privacy Policy. We may collect information, including personal information, from you if you choose to provide that information. For example, you may provide us with your name and contact information (such as an email address). You also may choose to provide personal information about yourself when you participate in surveys, forums or discussions, or other Interactive Services on the Sites. Please be aware that information you post may be viewed or captured by anyone who visits the Sites, therefore you should avoid posting sensitive or personal information that you would not want to be available to the public. We, and our third-party partners and service providers, may use the information gathered on the “What can I cook” site for various purposes. For example, if you were to contact us with a question and provide your email address, we would use the email address you provided to respond to your inquiry.
-<br></br>
-<br></br>
-<br></br>
-
-5. Dotdash Intellectual Property
-<br></br>
-<br></br>
-You agree not to modify, reproduce, distribute, disseminate, sell, publish, broadcast or circulate any such material from “What can I cook” site. We grants you a personal, non-exclusive, non-transferable, revocable license to use the site and any materials on the site for non-commercial purposes subject to these Terms of Use.
-
-<br></br>
-<br></br>
-<br></br>
-
-6. Reliance on Information on Site
-<br></br>
-<br></br>
-We have no obligation to, and you should not expect us to, review content on “What can I cook” site, including User Contributions.
-We does not represent or guarantee that any contributor has achieved any particular level of expertise or knowledge or has any specific qualifications or credentials, without limitation, as to the subject matter to which their contributions relate. To the extent we refer to each of these contributors as an expert on the site, you must understand that we rely on the information they provide us and we are not obligated to independently verify or attempt to confirm any information they provide, nor their qualifications or credentials. We are also not obligated to monitor or independently research or verify any content they contribute. We cannot and does not represent or warrant the accuracy, completeness or truthfulness of the qualifications or credentials of any contributor, nor of any other users of the sites.
-Please do not rely on our site’s content, including User Contributions and content from our independent contractor contributors. Content is provided for general information purposes only and can never take into account your unique, personal circumstances and needs. You acknowledge and agree that any reliance or actions you take in violation of your agreement with us shall be at your sole and exclusive risk and we shall have no responsibility or liability to you whatsoever. You also acknowledge and agree that communications on or through our site, whether with content providers or other users, are at your own risk and are not covered by any privilege or confidentiality obligation that might apply if you were to obtain your own professional advice.
-<br></br>
-<br></br>
-<br></br>
-
-7. Prohibited Uses of the Sites
-<br></br>
-<br></br>
-You may use the “What can I cook” site only for lawful purposes and in accordance with these Terms of Use. You agree not to use the site:
-In any way that violates any applicable federal, state, local or international law or regulation.
-For the purpose of exploiting, harming or attempting to exploit or harm minors in any way by exposing them to inappropriate content, asking for personally identifiable information or otherwise.
-To transmit, or procure the sending of, any advertising or promotional material, including any “junk mail”, “chain letter” or “spam” or any other similar solicitation.
-To impersonate or attempt to impersonate another user or any other person or entity (including, without limitation, by using email addresses or screen names associated with any of the foregoing).
-To engage in any other conduct that restricts or inhibits anyone’s use or enjoyment of the site, or which, as determined by us, may harm us or users of the site or expose them to liability.
-Additionally, you agree not to:
-“Scrape” or disaggregate data from the “What can I cook?” site (whether by manual or automated means) for any commercial, marketing, or data compiling or enhancing purpose.
-Introduce any viruses, Trojan horses, worms, logic bombs or other material or mechanism which is malicious or technologically harmful.
-Attempt to gain unauthorized access to, interfere with, damage or disrupt any parts of the “What can I cook?” site, the server on which the site is stored, or any server, computer or database connected to the site.
-Otherwise attempt to interfere with the proper working of the site.
-<br></br>
-<br></br>
-<br></br>
-
-8. Content That You Make Available on the Site
-<br></br>
-<br></br>
-If you voluntarily disclose personal information (e.g., user name, email address) on the “What can I cook” site, that information can be viewed in search engines, collected and used by others and may result in unsolicited contact from other parties. We strongly advise that you not post any personal or other sensitive information on our site.
-Any User Contribution you post to the site will be considered non-confidential and non-proprietary. By providing any User Contribution on the site, you grant us and our affiliates and service providers, and each of their and our respective licensees, successors and assigns the perpetual right to use, reproduce, modify, perform, display, distribute and otherwise disclose to third parties any such material for any purpose. 
-You represent and warrant that:
-You own or control all rights in and to the User Contributions and have the right to grant the license granted above to us and our affiliates and service providers, and each of their and our respective licensees, successors and assigns.
-All of your User Contributions do and will comply with these Terms of Use.
-You understand and acknowledge that you are responsible for any User Contributions you submit or contribute, and you, not we, have full responsibility for such content, including its legality, reliability, accuracy and appropriateness. We are not responsible, or liable to any third party, for the content or accuracy of any User Contributions posted by you or any other user of the Sites.
-<br></br>
-<br></br>
-Monitoring and Enforcement; Termination
-We have the right to:
-Remove or refuse to post any User Contributions for any or no reason in our sole discretion.
-Take any action with respect to any User Contribution that we deem necessary or appropriate in our sole discretion, including if we believe that such User Contribution violates the Terms of Use, including the content standards below, infringes any intellectual property right or other right of any person or entity, threatens the personal safety of users of the “What can I cook?” site or the public.
-Disclose your identity or other information about you to any third party who claims that material posted by you violates their rights, including their intellectual property rights or their right to privacy.
-Take appropriate legal action, including without limitation, referral to law enforcement, for any illegal or unauthorized use of the Sites.
-Terminate or suspend your access to all or part of the site for any or no reason, including without limitation, any violation of these Terms of Use.
-We cannot and do not undertake to review all material before it is posted on the Sites, and cannot ensure prompt removal of objectionable material after it has been posted. Accordingly, we assume no liability for any action or inaction regarding transmissions, communications or content provided by any user or third party. We have no liability or responsibility to anyone for performance or nonperformance of the activities described in this section.
-<br></br>
-<br></br>
-Content Standards
-These content standards apply to any and all User Contributions and use of Interactive Services. User Contributions must in their entirety comply with all applicable federal, state, local and international laws and regulations. Without limiting the foregoing, User Contributions must not:
-Contain any material which is defamatory, obscene, indecent, abusive, offensive, harassing, violent, hateful, inflammatory or otherwise objectionable.
-Promote sexually explicit or pornographic material, violence, or discrimination based on race, sex, religion, nationality, disability, sexual orientation or age.
-Infringe any patent, trademark, trade secret, copyright or other intellectual property or other rights of any other person.
-Violate the legal rights (including the rights of publicity and privacy) of others or contain any material that could give rise to any civil or criminal liability under applicable laws or regulations or that otherwise may be in conflict with these Terms of Use and our Privacy Policy.
-Be likely to deceive any person.
-Promote any illegal activity, or advocate, promote or assist any unlawful act.
-Cause annoyance, inconvenience or needless anxiety or be likely to upset, harass, embarrass, alarm or annoy any other person.
-Impersonate any person, or misrepresent your identity or affiliation with any person or organization.
-Promote commercial activities or sales, such as contests, sweepstakes and other sales promotions, barter or advertising.
-Give the impression that they emanate from or are endorsed by us or any other person or entity.
-<br></br>
-<br></br>
-<br></br>
-
-9. Your Indemnification of Us
-<br></br>
-<br></br>
-You agree to indemnify, defend and hold harmless us from and against any and all liability and costs, in connection with any claim arising out of (a) any User Contributions, or (b) breach by you or any user of your account of these Terms of Use or any representations, warranties and covenants contained in these Terms of Use.
-<br></br>
-<br></br>
-<br></br>
-
-10. WARRANTY DISCLAIMER
-<br></br>
-<br></br>
-THE SITES ARE PROVIDED ON AN “AS IS” BASIS WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO WARRANTIES OF TITLE OR IMPLIED WARRANTIES OF MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE, OTHER THAN THOSE WARRANTIES WHICH ARE IMPOSED BY AND INCAPABLE OF EXCLUSION, RESTRICTION OR MODIFICATION UNDER THE LAWS APPLICABLE TO THESE TERMS OF USE. WE DO NOT ENDORSE AND ARE NOT RESPONSIBLE FOR THE ACCURACY OR RELIABILITY OF ANY OPINION, ADVICE OR STATEMENT ON THE SITE. THE INFORMATION, FACTS, AND OPINIONS PROVIDED ARE NO SUBSTITUTE FOR PROFESSIONAL ADVICE.
-<br></br>
-<br></br>
-<br></br>
-
-11. LIABILITY DISCLAIMER
-<br></br>
-<br></br>
-YOUR USE OF THE SITES IS AT YOUR OWN RISK. WE ARE NOT SHALL BE LIABLE TO YOU OR ANYONE ELSE FOR ANY LOSS OR INJURY OR ANY DIRECT, INDIRECT, INCIDENTAL, CONSEQUENTIAL, SPECIAL, PUNITIVE OR SIMILAR DAMAGES ARISING OUT OF YOUR ACCESS OR USE OF, OR YOUR INABILITY TO ACCESS OR USE, THE SITES AND THE INFORMATION AVAILABLE ON THE SITE OR ARISING OUT OF ANY ACTION TAKEN IN RESPONSE TO OR AS A RESULT OF ANY INFORMATION AVAILABLE ON THE SITES. YOU HEREBY WAIVE ANY AND ALL CLAIMS AGAINST US AND ITS SUBSIDIARIES, DIVISIONS, AFFILIATES, AGENTS, REPRESENTATIVES AND LICENSORS (INCLUDING OUR INDEPENDENT CONTRACTOR CONTRIBUTORS) ARISING OUT OF YOUR USE OF THE SITE AND THE INFORMATION AVAILABLE THEREON.
-<br></br>
-<br></br>
-<br></br>
-
-12. Third-Party Links, Advertisements, Websites and Content
-<br></br>
-<br></br>
-We do not review or monitor any websites, advertisements, or other media linked to or available through the Sites and are not responsible for the content of any such third-party advertisements or linked websites.
-</p>
-    </div>
-       </React.Fragment>
-    );
-};
-
-export default TermsOfUse;
