+++ src/app/terms/page.js
@@ -1,11 +1,4 @@
-"use client";
-import "primereact/resources/primereact.min.css";
-import "primeicons/primeicons.css";
-import "primeflex/primeflex.css";
-import "../page.css";
-
-
-import { TermsOfUse } from "../components/terms/terms";
+import { TermsOfUse } from "./terms";
 
 export default function Home() {
   return (
